feat: Folder Explorer as an independent component (#2529)
### TL;DR

This is a task to decouple folder-explorer from storage-list so that folder-explorer can be called and used from anywhere in React Root.

### What changed?

The folder explorer that was dependent on `backend-ai-storage-list` has been separated and used inside `LegacyFolderExplorer`, which can be called and used from any part of react root.

The API response to get files of a vfolder exists files(for legacy compatibility) and items. In explorer, I changed to using items.

Currently, you can use the folder ID passed as a query parameter by passing it as a prop.

### How to test?

1. navigate to the Data & Storage tab.
2. create a folder or open an existing folder.
3. Check the folder explorer modal.

The `LegacyFolderExplorer` provides the same functionality as before. I'd appreciate it if you could check if there are any changes I missed compared to the existing modal.

### Why make this change?

To improve the usability and functionality of the modal and folder explorer components in the application.

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after
