+++ resources/i18n/fr.json
@@ -631,7 +631,6 @@
       "Delete": "Effacer",
       "FolderOptionUpdate": "Option de mise à jour du dossier",
       "Leave": "Quitter",
-      "RenameAFolder": "Renommer un dossier",
       "TypeNewFolderName": "Tapez un nouveau nom de dossier",
       "FolderCreated": "Dossier créé",
       "FolderCloned": "Dossier cloné",
@@ -711,7 +710,6 @@
       "FileExtensionChanged": "Souhaitez-vous modifier l'extension de fichier ?",
       "KeepFileExtension": "Garder",
       "UseNewFileExtension": "Utiliser",
-      "RemoveFileExtension": "supprimer l'extension de fichier",
       "ExecutingFileBrowser": "Exécution du navigateur de fichiers...",
       "ExecuteFileBrowser": "Exécuter le navigateur de fichiers",
       "NotEnoughResourceForFileBrowserSession": "Pas assez de ressources (cpu : 1 Core, mem : 0,5 Go) pour créer la session pour le navigateur de fichiers. veuillez vérifier les ressources disponibles.",
@@ -727,7 +725,10 @@
       "FolderAlreadyExists": "Le dossier sur le point d'être téléchargé existe déjà.",
       "NumberOfSFTPSessionsExceededTitle": "Limite atteinte pour le nombre de sessions de téléchargement en cours",
       "NumberOfSFTPSessionsExceededBody": "Vous exécutez toutes les sessions de téléchargement que vous êtes autorisé à créer. Veuillez mettre fin aux sessions de téléchargement inutilisées avant de commencer une nouvelle session.",
-      "DownloadNotAllowed": "Le téléchargement de fichiers/dossiers n'est pas autorisé dans ce dossier."
+      "DownloadNotAllowed": "Le téléchargement de fichiers/dossiers n'est pas autorisé dans ce dossier.",
+      "RenameAFolder": "Renommer le dossier",
+      "Filename": "Nom de fichier",
+      "RemoveFileExtension": "Retirer "
     },
     "invitation": {
       "NoValidEmails": "Aucun e-mail valide n'a été saisi",
