+++ resources/i18n/zh-CN.json
@@ -631,7 +631,6 @@
       "Delete": "删除",
       "FolderOptionUpdate": "更新文件夹选项",
       "Leave": "离开",
-      "RenameAFolder": "重命名文件夹",
       "TypeNewFolderName": "输入新文件夹名称",
       "FolderCreated": "文件夹已创建",
       "FolderCloned": "文件夹已克隆",
@@ -711,7 +710,6 @@
       "FileExtensionChanged": "您想更改文件扩展名吗？",
       "KeepFileExtension": "保持",
       "UseNewFileExtension": "用",
-      "RemoveFileExtension": "删除文件扩展名",
       "ExecutingFileBrowser": "正在执行文件浏览器...",
       "ExecuteFileBrowser": "执行文件浏览器",
       "NotEnoughResourceForFileBrowserSession": "没有足够的资源（cpu: 1 Core, mem: 0.5GB）来为文件浏览器创建会话。请检查可用资源。",
@@ -727,7 +725,10 @@
       "StartingSSH/SFTPSession": "开始 SFTP 会话...",
       "NumberOfSFTPSessionsExceededTitle": "运行中的上传会话次数已达上限",
       "NumberOfSFTPSessionsExceededBody": "您正在运行允许创建的所有可用上传会话。请在启动新会话前终止未使用的上传会话。",
-      "DownloadNotAllowed": "不允许在此文件夹中下载文件/文件夹。"
+      "DownloadNotAllowed": "不允许在此文件夹中下载文件/文件夹。",
+      "RenameAFolder": "重命名文件夹",
+      "Filename": "文件名",
+      "RemoveFileExtension": "消除 "
     },
     "invitation": {
       "NoValidEmails": "没有输入有效的电子邮件",
