+++ src/components/backend-ai-storage-list.ts
@@ -3,7 +3,6 @@
  Copyright (c) 2015-2024 Lablup Inc. All rights reserved.
  */
 import { navigate } from '../backend-ai-app';
-import tus from '../lib/tus';
 import '../plastics/lablup-shields/lablup-shields';
 import {
   IronFlex,
@@ -23,7 +22,6 @@ import BackendAiSessionLauncher from './backend-ai-session-launcher';
 import './lablup-grid-sort-filter-column';
 import './lablup-loading-spinner';
 import LablupLoadingSpinner from './lablup-loading-spinner';
-import { Button } from '@material/mwc-button';
 import '@material/mwc-formfield';
 import '@material/mwc-icon-button';
 import '@material/mwc-list';
@@ -63,14 +61,6 @@ interface inviteeData {
   vfolder_name: string;
 }
 
-interface fileData {
-  id: string;
-  progress: number;
-  caption: string;
-  error: boolean;
-  complete: boolean;
-}
-
 type VFolderOperationStatus =
   | 'ready'
   | 'performing'
@@ -106,7 +96,6 @@ type DeadVFolderStatus =
 
 @customElement('backend-ai-storage-list')
 export default class BackendAiStorageList extends BackendAIPage {
-  @property({ type: Number }) _APIMajorVersion = 5;
   @property({ type: String }) storageType = 'general';
   @property({ type: Array }) folders = [];
   @property({ type: Object }) folderInfo = Object();
@@ -119,67 +108,36 @@ export default class BackendAiStorageList extends BackendAIPage {
   @property({ type: String }) deleteFolderName = '';
   @property({ type: String }) deleteFolderID = '';
   @property({ type: String }) leaveFolderName = '';
-  @property({ type: Object }) explorer = Object();
-  @property({ type: Array }) explorerFiles = [];
   @property({ type: String }) existingFile = '';
   @property({ type: Array }) invitees: inviteeData[] = [];
   @property({ type: String }) selectedFolder = '';
   @property({ type: String }) selectedFolderType = '';
-  @property({ type: String }) downloadURL = '';
-  @property({ type: Array }) uploadFiles: fileData[] = [];
-  @property({ type: Object }) currentUploadFile = Object();
-  @property({ type: Array }) fileUploadQueue = [];
-  @property({ type: Number }) fileUploadCount = 0;
-  @property({ type: Number }) concurrentFileUploadLimit = 2;
-  @property({ type: String }) vhost = '';
-  @property({ type: Array }) vhosts = [];
-  @property({ type: Array }) allowedGroups = [];
   @property({ type: Object }) indicator = Object();
   @property({ type: Object }) notification = Object();
   @property({ type: String }) listCondition: StatusCondition = 'loading';
   @property({ type: Array }) allowed_folder_type = [];
-  @property({ type: Boolean }) uploadFilesExist = false;
   @property({ type: Object }) _boundIndexRenderer = Object();
   @property({ type: Object }) _boundTypeRenderer = Object();
   @property({ type: Object }) _boundFolderListRenderer = Object();
   @property({ type: Object }) _boundControlFolderListRenderer = Object();
   @property({ type: Object }) _boundTrashBinControlFolderListRenderer =
     Object();
-  @property({ type: Object }) _boundControlFileListRenderer = Object();
   @property({ type: Object }) _boundPermissionViewRenderer = Object();
   @property({ type: Object }) _boundOwnerRenderer = Object();
-  @property({ type: Object }) _boundFileNameRenderer = Object();
-  @property({ type: Object }) _boundCreatedTimeRenderer = Object();
-  @property({ type: Object }) _boundSizeRenderer = Object();
   @property({ type: Object }) _boundPermissionRenderer = Object();
   @property({ type: Object }) _boundCloneableRenderer = Object();
   @property({ type: Object }) _boundQuotaRenderer = Object();
-  @property({ type: Object }) _boundUploadListRenderer = Object();
-  @property({ type: Object }) _boundUploadProgressRenderer = Object();
   @property({ type: Object }) _boundInviteeInfoRenderer = Object();
   @property({ type: Object }) _boundIDRenderer = Object();
   @property({ type: Object }) _boundStatusRenderer = Object();
-  @property({ type: Boolean }) _uploadFlag = true;
   @property({ type: Boolean }) _folderRefreshing = false;
   @property({ type: Number }) lastQueryTime = 0;
-  @property({ type: Boolean }) isWritable = false;
   @property({ type: Object }) permissions = {
     rw: 'Read-Write',
     ro: 'Read-Only',
     wd: 'Delete',
   };
-  @property({ type: Number }) _maxFileUploadSize = -1;
   @property({ type: Number }) selectAreaHeight;
-  @property({ type: String }) oldFileExtension = '';
-  @property({ type: String }) newFileExtension = '';
-  @property({ type: Boolean }) is_dir = false;
-  @property({ type: Boolean }) _isDirectorySizeVisible = true;
-  @property({ type: Number }) minimumResource = {
-    cpu: 1,
-    mem: 0.5,
-  };
-  @property({ type: Array }) filebrowserSupportedImages = [];
-  @property({ type: Array }) systemRoleSupportedImages = [];
   @property({ type: Object }) volumeInfo = Object();
   @property({ type: Array }) quotaSupportStorageBackends = [
     'xfs',
@@ -213,25 +171,14 @@ export default class BackendAiStorageList extends BackendAIPage {
   @query('#list-status') private _listStatus!: BackendAIListStatus;
   @query('#modify-folder-quota') modifyFolderQuotaInput!: TextField;
   @query('#modify-folder-quota-unit') modifyFolderQuotaUnitSelect!: Select;
-  @query('#file-list-grid') fileListGrid!: VaadinGrid;
   @query('#folder-list-grid') folderListGrid!: VaadinGrid;
-  @query('#mkdir-name') mkdirNameInput!: TextField;
   @query('#delete-folder-name') deleteFolderNameInput!: TextField;
   @query('#delete-from-trash-bin-name-input')
   deleteFromTrashBinNameInput!: TextField;
   @query('#new-folder-name') newFolderNameInput!: TextField;
-  @query('#new-file-name') newFileNameInput!: TextField;
   @query('#leave-folder-name') leaveFolderNameInput!: TextField;
   @query('#update-folder-permission') updateFolderPermissionSelect!: Select;
   @query('#update-folder-cloneable') updateFolderCloneableSwitch!: Switch;
-  @query('#rename-file-dialog') renameFileDialog!: BackendAIDialog;
-  @query('#delete-file-dialog') deleteFileDialog!: BackendAIDialog;
-  @query('#filebrowser-notification-dialog')
-  fileBrowserNotificationDialog!: BackendAIDialog;
-  @query('#file-extension-change-dialog')
-  fileExtensionChangeDialog!: BackendAIDialog;
-  @query('#folder-explorer-dialog') folderExplorerDialog!: BackendAIDialog;
-  @query('#download-file-dialog') downloadFileDialog!: BackendAIDialog;
   @query('#modify-permission-dialog') modifyPermissionDialog!: BackendAIDialog;
   @query('#share-folder-dialog') shareFolderDialog!: BackendAIDialog;
   @query('#session-launcher') sessionLauncher!: BackendAiSessionLauncher;
@@ -244,19 +191,12 @@ export default class BackendAiStorageList extends BackendAIPage {
       this.controlFolderListRenderer.bind(this);
     this._boundTrashBinControlFolderListRenderer =
       this.trashBinControlFolderListRenderer.bind(this);
-    this._boundControlFileListRenderer =
-      this.controlFileListRenderer.bind(this);
     this._boundPermissionViewRenderer = this.permissionViewRenderer.bind(this);
     this._boundCloneableRenderer = this.CloneableRenderer.bind(this);
     this._boundOwnerRenderer = this.OwnerRenderer.bind(this);
-    this._boundFileNameRenderer = this.fileNameRenderer.bind(this);
-    this._boundCreatedTimeRenderer = this.createdTimeRenderer.bind(this);
-    this._boundSizeRenderer = this.sizeRenderer.bind(this);
     this._boundPermissionRenderer = this.permissionRenderer.bind(this);
     this._boundFolderListRenderer = this.folderListRenderer.bind(this);
     this._boundQuotaRenderer = this.quotaRenderer.bind(this);
-    this._boundUploadListRenderer = this.uploadListRenderer.bind(this);
-    this._boundUploadProgressRenderer = this.uploadProgressRenderer.bind(this);
     this._boundInviteeInfoRenderer = this.inviteeInfoRenderer.bind(this);
     this._boundIDRenderer = this.iDRenderer.bind(this);
     this._boundStatusRenderer = this.statusRenderer.bind(this);
@@ -350,73 +290,12 @@ export default class BackendAiStorageList extends BackendAIPage {
           --mdc-theme-secondary: var(--general-checkbox-color);
         }
 
-        #folder-explorer-dialog {
-          width: calc(100% - 250px); /* 250px is width for drawer menu */
-          --component-height: calc(100vh - 200px); /* calc(100vh - 170px); */
-          right: 0;
-          top: 0;
-          margin: 170px 0 0 0;
-        }
-
-        #folder-explorer-dialog.mini_ui {
-          width: calc(
-            100% - 88px
-          ); /* 88px is width for mini-ui icon of drawer menu */
-        }
-
-        /* #folder-explorer-dialog vaadin-grid vaadin-grid-column {
-          height: 32px !important;
-        }*/
-
-        #folder-explorer-dialog vaadin-grid mwc-icon-button {
-          --mdc-icon-size: 24px;
-          --mdc-icon-button-size: 28px;
-          background-color: transparent;
-        }
-
-        #filebrowser-notification-dialog {
-          --component-width: 350px;
-        }
-
         vaadin-text-field {
           --vaadin-text-field-default-width: auto;
         }
 
         vaadin-grid-cell-content {
           overflow: visible;
-        }
-
-        div.breadcrumb {
-          color: #637282;
-          font-size: 1em;
-          margin-bottom: 10px;
-          margin-left: 20px;
-        }
-
-        div.breadcrumb span:first-child {
-          display: none;
-        }
-
-        .breadcrumb li:before {
-          padding: 3px;
-          transform: rotate(-45deg) translateY(-2px);
-          transition: color ease-in 0.2s;
-          border: solid;
-          border-width: 0 2px 2px 0;
-          border-color: var(--token-colorBorder, #242424);
-          margin-right: 10px;
-          content: '';
-          display: inline-block;
-        }
-
-        .breadcrumb li {
-          display: inline-block;
-          font-size: 16px;
-        }
-
-        .breadcrumb mwc-icon-button {
-          --mdc-icon-size: 20px;
-          --mdc-icon-button-size: 22px;
         }
 
         mwc-textfield {
@@ -445,24 +324,6 @@ export default class BackendAiStorageList extends BackendAIPage {
         div#upload {
           margin: 0;
           padding: 0;
-        }
-
-        div#dropzone {
-          display: none;
-          position: absolute;
-          top: 0;
-          height: 100%;
-          width: 100%;
-          z-index: 10;
-        }
-
-        div#dropzone,
-        div#dropzone p {
-          margin: 0;
-          padding: 0;
-          width: 100%;
-          background: rgba(211, 211, 211, 0.5);
-          text-align: center;
         }
 
         .progress {
@@ -524,32 +385,7 @@ export default class BackendAiStorageList extends BackendAIPage {
           filter: grayscale(1);
         }
 
-        img#filebrowser-img,
-        img#ssh-img {
-          width: 18px;
-          margin: 15px 10px;
-        }
-
-        @media screen and (max-width: 700px) {
-          #folder-explorer-dialog,
-          #folder-explorer-dialog.mini_ui {
-            min-width: 410px;
-            --component-width: 100%;
-            width: 100%;
-            position: absolute;
-            margin-left: auto;
-            margin-right: auto;
-            left: 0px;
-            right: 0px;
-          }
-        }
-
         @media screen and (max-width: 750px) {
-          #folder-explorer-dialog,
-          #folder-explorer-dialog.mini_ui {
-            --component-width: auto;
-          }
-
           mwc-button {
             width: auto;
           }
@@ -560,30 +396,8 @@ export default class BackendAiStorageList extends BackendAIPage {
             --component-min-width: 100%;
           }
         }
-
-        @media screen and (min-width: 900px) {
-          #folder-explorer-dialog,
-          #folder-explorer-dialog.mini_ui {
-            --component-width: calc(100% - 45px); /* calc(100% - 30px); */
-          }
-        }
       `,
     ];
-  }
-
-  _toggleFileListCheckbox() {
-    const buttons = this.shadowRoot?.querySelectorAll<Button>(
-      '.multiple-action-buttons',
-    ) as NodeListOf<Button>;
-    if (this.fileListGrid.selectedItems.length > 0) {
-      [].forEach.call(buttons, (e: HTMLElement) => {
-        e.style.display = 'block';
-      });
-    } else {
-      [].forEach.call(buttons, (e: HTMLElement) => {
-        e.style.display = 'none';
-      });
-    }
   }
 
   /**
@@ -1131,263 +945,6 @@ export default class BackendAiStorageList extends BackendAIPage {
           </mwc-list>
         </div>
       </backend-ai-dialog>
-      <backend-ai-dialog
-        id="folder-explorer-dialog"
-        class="folder-explorer"
-        narrowLayout
-        scrimClickAction
-        @dialog-closed=${() => {
-          this.triggerCloseFilebrowserToReact();
-        }}
-      >
-        <span slot="title" style="margin-right:1rem;">${this.explorer.id}</span>
-        <div
-          slot="action"
-          class="horizontal layout space-between folder-action-buttons center"
-        >
-          <div class="flex"></div>
-          ${this.isWritable
-            ? html`
-                <mwc-button
-                  outlined
-                  class="multiple-action-buttons fg red"
-                  icon="delete"
-                  @click="${() => this._openDeleteMultipleFileDialog()}"
-                  style="display:none;"
-                >
-                  <span>${_t('data.explorer.Delete')}</span>
-                </mwc-button>
-                <div id="add-btn-cover">
-                  <mwc-button
-                    id="add-btn"
-                    icon="upload_file"
-                    ?disabled=${!this.isWritable}
-                    @click="${(e) => this._uploadBtnClick(e)}"
-                  >
-                    <span>${_t('data.explorer.UploadFiles')}</span>
-                  </mwc-button>
-                </div>
-                <div>
-                  <mwc-button
-                    id="add-folder-btn"
-                    icon="drive_folder_upload"
-                    ?disabled=${!this.isWritable}
-                    @click="${(e) => this._uploadBtnClick(e)}"
-                  >
-                    <span>${_t('data.explorer.UploadFolder')}</span>
-                  </mwc-button>
-                </div>
-                <div id="mkdir-cover">
-                  <mwc-button
-                    id="mkdir"
-                    class="tooltip"
-                    icon="create_new_folder"
-                    ?disabled=${!this.isWritable}
-                    @click="${() => this._mkdirDialog()}"
-                  >
-                    <span>${_t('data.explorer.NewFolder')}</span>
-                  </mwc-button>
-                </div>
-              `
-            : html`
-                <mwc-button id="readonly-btn" disabled>
-                  <span>${_t('data.explorer.ReadonlyFolder')}</span>
-                </mwc-button>
-              `}
-          <div id="filebrowser-btn-cover">
-            <mwc-button
-              id="filebrowser-btn"
-              @click="${() => this._executeFileBrowser()}"
-            >
-              <img
-                id="filebrowser-img"
-                alt="File Browser"
-                src="./resources/icons/filebrowser.svg"
-              />
-              <span>${_t('data.explorer.ExecuteFileBrowser')}</span>
-            </mwc-button>
-          </div>
-          <div>
-            <mwc-button
-              id="ssh-btn"
-              title="SSH / SFTP"
-              @click="${() => this._executeSSHProxyAgent()}"
-            >
-              <img
-                id="ssh-img"
-                alt="SSH / SFTP"
-                src="/resources/icons/sftp.png"
-              />
-              <span>${_t('data.explorer.RunSSH/SFTPserver')}</span>
-            </mwc-button>
-          </div>
-        </div>
-        <div slot="content">
-          <div class="breadcrumb">
-            ${this.explorer.breadcrumb
-              ? html`
-                  <ul>
-                    ${this.explorer.breadcrumb.map(
-                      (item) => html`
-                        <li>
-                          ${item === '.'
-                            ? html`
-                                <mwc-icon-button
-                                  icon="folder_open"
-                                  dest="${item}"
-                                  @click="${(e) => this._gotoFolder(e)}"
-                                ></mwc-icon-button>
-                              `
-                            : html`
-                                <a
-                                  outlined
-                                  class="goto"
-                                  path="item"
-                                  @click="${(e) => this._gotoFolder(e)}"
-                                  dest="${item}"
-                                >
-                                  ${item}
-                                </a>
-                              `}
-                        </li>
-                      `,
-                    )}
-                  </ul>
-                `
-              : html``}
-          </div>
-          <div id="dropzone"><p>drag</p></div>
-          <input
-            type="file"
-            id="fileInput"
-            @change="${(e) => this._uploadInputChange(e)}"
-            hidden
-            multiple
-          />
-          <input
-            type="file"
-            id="folderInput"
-            @change="${(e) => this._uploadInputChange(e)}"
-            hidden
-            webkitdirectory
-            mozdirectory
-            directory
-            multiple
-          />
-          ${this.uploadFilesExist
-            ? html`
-                <div class="horizontal layout start-justified">
-                  <mwc-button
-                    icon="cancel"
-                    id="cancel_upload"
-                    @click="${() => this._cancelUpload()}"
-                  >
-                    ${_t('data.explorer.StopUploading')}
-                  </mwc-button>
-                </div>
-                <div class="horizontal layout center progress-item flex">
-                  ${this.currentUploadFile?.complete
-                    ? html`
-                        <mwc-icon>check</mwc-icon>
-                      `
-                    : html``}
-                  <div
-                    class="vertical layout progress-item"
-                    style="width:100%;"
-                  >
-                    <span>${this.currentUploadFile?.name}</span>
-                    <vaadin-progress-bar
-                      value="${this.currentUploadFile?.progress}"
-                    ></vaadin-progress-bar>
-                    <span>${this.currentUploadFile?.caption}</span>
-                  </div>
-                </div>
-                <!-- <vaadin-grid class="progress" theme="row-stripes compact" aria-label="uploadFiles" .items="${this
-                  .uploadFiles}" height-by-rows>
-            <vaadin-grid-column width="100px" flex-grow="0" .renderer="${this
-                  ._boundUploadListRenderer}"></vaadin-grid-column>
-            <vaadin-grid-column .renderer="${this
-                  ._boundUploadProgressRenderer}"></vaadin-grid-column>
-          </vaadin-grid> -->
-              `
-            : html``}
-          <vaadin-grid
-            id="file-list-grid"
-            class="explorer"
-            theme="row-stripes compact dark"
-            aria-label="Explorer"
-            .items="${this.explorerFiles}"
-          >
-            <vaadin-grid-selection-column
-              auto-select
-            ></vaadin-grid-selection-column>
-            <vaadin-grid-column
-              width="40px"
-              flex-grow="0"
-              resizable
-              header="#"
-              .renderer="${this._boundIndexRenderer}"
-            ></vaadin-grid-column>
-            <vaadin-grid-sort-column
-              flex-grow="2"
-              resizable
-              header="${_t('data.explorer.Name')}"
-              path="filename"
-              .renderer="${this._boundFileNameRenderer}"
-            ></vaadin-grid-sort-column>
-            <vaadin-grid-sort-column
-              flex-grow="2"
-              resizable
-              header="${_t('data.explorer.Created')}"
-              path="ctime"
-              .renderer="${this._boundCreatedTimeRenderer}"
-            ></vaadin-grid-sort-column>
-            <vaadin-grid-sort-column
-              auto-width
-              resizable
-              header="${_t('data.explorer.Size')}"
-              path="size"
-              .renderer="${this._boundSizeRenderer}"
-            ></vaadin-grid-sort-column>
-            <vaadin-grid-column
-              resizable
-              auto-width
-              header="${_t('data.explorer.Actions')}"
-              .renderer="${this._boundControlFileListRenderer}"
-            ></vaadin-grid-column>
-          </vaadin-grid>
-        </div>
-      </backend-ai-dialog>
-      <backend-ai-dialog id="mkdir-dialog" fixed backdrop>
-        <span slot="title">${_t('data.explorer.CreateANewFolder')}</span>
-        <div slot="content">
-          <mwc-textfield
-            id="mkdir-name"
-            label="${_t('data.explorer.Foldername')}"
-            @change="${() => this._validatePathName()}"
-            required
-            maxLength="255"
-            placeholder="${_text('maxLength.255chars')}"
-            validationMessage="${_text('data.explorer.ValueRequired')}"
-          ></mwc-textfield>
-          <br />
-        </div>
-        <div
-          slot="footer"
-          class="horizontal center-justified flex layout distancing"
-        >
-          <mwc-button
-            icon="rowing"
-            unelevated
-            fullwidth
-            type="submit"
-            id="mkdir-btn"
-            @click="${(e) => this._mkdir(e)}"
-          >
-            ${_t('button.Create')}
-          </mwc-button>
-        </div>
-      </backend-ai-dialog>
       <backend-ai-dialog id="share-folder-dialog" fixed backdrop>
         <span slot="title">${_t('data.explorer.ShareFolder')}</span>
         <div slot="content" role="listbox" style="margin: 0;width:100%;">
@@ -1482,148 +1039,6 @@ export default class BackendAiStorageList extends BackendAIPage {
           </mwc-button>
         </div>
       </backend-ai-dialog>
-      <backend-ai-dialog id="rename-file-dialog" fixed backdrop blockscrolling>
-        <span slot="title">${_t('data.explorer.RenameAFile')}</span>
-        <div slot="content">
-          <mwc-textfield
-            class="red"
-            id="new-file-name"
-            label="${_t('data.explorer.NewFileName')}"
-            required
-            @change="${() => this._validateExistingFileName()}"
-            auto-validate
-            style="width:320px;"
-            maxLength="255"
-            placeholder="${_text('maxLength.255chars')}"
-            autoFocus
-          ></mwc-textfield>
-          <div id="old-file-name" style="padding-left:15px;height:2.5em;"></div>
-        </div>
-        <div slot="footer" class="horizontal center-justified flex layout">
-          <mwc-button
-            icon="edit"
-            fullwidth
-            type="button"
-            id="rename-file-button"
-            unelevated
-            @click="${() => this._compareFileExtension()}"
-          >
-            ${_t('data.explorer.RenameAFile')}
-          </mwc-button>
-        </div>
-      </backend-ai-dialog>
-      <backend-ai-dialog id="delete-file-dialog" fixed backdrop>
-        <span slot="title">${_t('dialog.title.LetsDouble-Check')}</span>
-        <div slot="content">
-          <p>
-            ${_t('dialog.warning.CannotBeUndone')}
-            ${_t('dialog.ask.DoYouWantToProceed')}
-          </p>
-        </div>
-        <div slot="footer" class="horizontal end-justified flex layout">
-          <mwc-button outlined @click="${(e) => this._hideDialog(e)}">
-            ${_t('button.Cancel')}
-          </mwc-button>
-          <mwc-button raised @click="${(e) => this._deleteFileWithCheck(e)}">
-            ${_t('button.Okay')}
-          </mwc-button>
-        </div>
-      </backend-ai-dialog>
-      <backend-ai-dialog id="download-file-dialog" fixed backdrop>
-        <span slot="title">${_t('data.explorer.DownloadFile')}</span>
-        <div slot="content">
-          <a href="${this.downloadURL}">
-            <mwc-button outlined>
-              ${_t('data.explorer.TouchToDownload')}
-            </mwc-button>
-          </a>
-        </div>
-        <div
-          slot="footer"
-          class="horizontal end-justified flex layout distancing"
-        >
-          <mwc-button @click="${(e) => this._hideDialog(e)}">
-            ${_t('button.Close')}
-          </mwc-button>
-        </div>
-      </backend-ai-dialog>
-      <backend-ai-dialog id="file-extension-change-dialog" fixed backdrop>
-        <span slot="title">${_t('dialog.title.LetsDouble-Check')}</span>
-        <div slot="content">
-          <p>${_t('data.explorer.FileExtensionChanged')}</p>
-        </div>
-        <div
-          slot="footer"
-          class="horizontal center-justified flex layout distancing"
-        >
-          <mwc-button
-            outlined
-            fullwidth
-            @click="${(e) => this._keepFileExtension()}"
-          >
-            ${globalThis.backendaioptions.get(
-              'language',
-              'default',
-              'general',
-            ) !== 'ko'
-              ? html`
-                  ${_text('data.explorer.KeepFileExtension') +
-                  this.oldFileExtension}
-                `
-              : html`
-                  ${this.oldFileExtension +
-                  _text('data.explorer.KeepFileExtension')}
-                `}
-          </mwc-button>
-          <mwc-button unelevated fullwidth @click="${() => this._renameFile()}">
-            ${globalThis.backendaioptions.get(
-              'language',
-              'default',
-              'general',
-            ) !== 'ko'
-              ? html`
-                  ${this.newFileExtension
-                    ? _text('data.explorer.UseNewFileExtension') +
-                      this.newFileExtension
-                    : _text('data.explorer.RemoveFileExtension')}
-                `
-              : html`
-                  ${this.newFileExtension
-                    ? this.newFileExtension +
-                      _text('data.explorer.UseNewFileExtension')
-                    : _text('data.explorer.RemoveFileExtension')}
-                `}
-          </mwc-button>
-        </div>
-      </backend-ai-dialog>
-      <backend-ai-dialog
-        id="filebrowser-notification-dialog"
-        fixed
-        backdrop
-        narrowLayout
-      >
-        <span slot="title">${_t('dialog.title.Notice')}</span>
-        <div slot="content" style="margin: 15px;">
-          <span>${_t('data.explorer.ReadOnlyFolderOnFileBrowser')}</span>
-        </div>
-        <div
-          slot="footer"
-          class="flex horizontal layout center justified"
-          style="margin: 15px 15px 15px 0px;"
-        >
-          <div class="horizontal layout start-justified center">
-            <mwc-checkbox
-              @change="${(e) => this._toggleShowFilebrowserNotification(e)}"
-            ></mwc-checkbox>
-            <span style="font-size:0.8rem;">
-              ${_text('dialog.hide.DonotShowThisAgain')}
-            </span>
-          </div>
-          <mwc-button unelevated @click="${(e) => this._hideDialog(e)}">
-            ${_t('button.Confirm')}
-          </mwc-button>
-        </div>
-      </backend-ai-dialog>
       <backend-ai-dialog id="delete-from-trash-bin-dialog" fixed backdrop>
         <span slot="title">${_t('dialog.title.DeleteForever')}</span>
         <div slot="content">
@@ -1657,11 +1072,6 @@ export default class BackendAiStorageList extends BackendAIPage {
   }
 
   firstUpdated() {
-    this._addEventListenerDropZone();
-    this._mkdir = this._mkdir.bind(this);
-    this.fileListGrid.addEventListener('selected-items-changed', () => {
-      this._toggleFileListCheckbox();
-    });
     this.indicator = globalThis.lablupIndicator;
     this.notification = globalThis.lablupNotification;
     const textfields = this.shadowRoot?.querySelectorAll(
@@ -1682,10 +1092,6 @@ export default class BackendAiStorageList extends BackendAIPage {
     document.addEventListener('backend-ai-group-changed', (e) =>
       this._refreshFolderList(true, 'group-changed'),
     );
-    document.addEventListener('backend-ai-ui-changed', (e) =>
-      this._refreshFolderUI(e),
-    );
-    this._refreshFolderUI({ detail: { 'mini-ui': globalThis.mini_ui } });
   }
 
   _modifySharedFolderPermissions() {
@@ -1859,48 +1265,6 @@ export default class BackendAiStorageList extends BackendAIPage {
         <div class="horizontal layout center center-justified">
           ${quotaIndicator}
         </div>
-      `,
-      root,
-    );
-  }
-
-  uploadListRenderer(root, column?, rowData?) {
-    render(
-      // language=HTML
-      html`
-        <vaadin-item class="progress-item">
-          <div>
-            ${rowData.item.complete
-              ? html`
-                  <mwc-icon>check</mwc-icon>
-                `
-              : html``}
-          </div>
-        </vaadin-item>
-      `,
-      root,
-    );
-  }
-
-  uploadProgressRenderer(root, column?, rowData?) {
-    render(
-      // language=HTML
-      html`
-        <vaadin-item>
-          <span>${rowData.item.name}</span>
-          ${rowData.item.complete
-            ? html``
-            : html`
-                <div>
-                  <vaadin-progress-bar
-                    value="${rowData.item.progress}"
-                  ></vaadin-progress-bar>
-                </div>
-                <div>
-                  <span>${rowData.item.caption}</span>
-                </div>
-              `}
-        </vaadin-item>
       `,
       root,
     );
@@ -2246,99 +1610,6 @@ export default class BackendAiStorageList extends BackendAIPage {
   }
 
   /**
-   * Render control file options - downloadFile, openRenameFileDialog, openDeleteFileDialog, etc.
-   *
-   * @param {Element} root - the row details content DOM element
-   * @param {Element} column - the column element that controls the state of the host element
-   * @param {Object} rowData - the object with the properties related with the rendered item
-   * */
-  controlFileListRenderer(root, column?, rowData?) {
-    render(
-      // language=HTML
-      html`
-        <div class="flex layout wrap">
-          <mwc-icon-button
-            id="${rowData.item.filename + '-download-btn'}"
-            class="tiny fg blue"
-            icon="cloud_download"
-            style="pointer-events: auto !important;"
-            ?disabled="${!this._isDownloadable(this.vhost)}"
-            filename="${rowData.item.filename}"
-            @click="${(e) => this._downloadFile(e, this._isDir(rowData.item))}"
-          ></mwc-icon-button>
-          ${!this._isDownloadable(this.vhost)
-            ? html`
-                <vaadin-tooltip
-                  for="${rowData.item.filename + '-download-btn'}"
-                  text="${_t('data.explorer.DownloadNotAllowed')}"
-                  position="top-start"
-                ></vaadin-tooltip>
-              `
-            : html``}
-          <mwc-icon-button
-            id="rename-btn"
-            ?disabled="${!this.isWritable}"
-            class="tiny fg green"
-            icon="edit"
-            required
-            filename="${rowData.item.filename}"
-            @click="${(e) =>
-              this._openRenameFileDialog(e, this._isDir(rowData.item))}"
-          ></mwc-icon-button>
-          <mwc-icon-button
-            id="delete-btn"
-            ?disabled="${!this.isWritable}"
-            class="tiny fg red"
-            icon="delete_forever"
-            filename="${rowData.item.filename}"
-            @click="${(e) => this._openDeleteFileDialog(e)}"
-          ></mwc-icon-button>
-        </div>
-      `,
-      root,
-    );
-  }
-
-  /**
-   * Render file name as rowData.item.filename.
-   *
-   * @param {Element} root - the row details content DOM element
-   * @param {Element} column - the column element that controls the state of the host element
-   * @param {Object} rowData - the object with the properties related with the rendered item
-   * */
-  fileNameRenderer(root, column?, rowData?) {
-    render(
-      html`
-        ${this._isDir(rowData.item)
-          ? html`
-              <div
-                class="indicator horizontal center layout"
-                name="${rowData.item.filename}"
-              >
-                <mwc-icon-button
-                  class="fg controls-running"
-                  icon="folder_open"
-                  name="${rowData.item.filename}"
-                  @click="${(e) => this._enqueueFolder(e)}"
-                ></mwc-icon-button>
-                ${rowData.item.filename}
-              </div>
-            `
-          : html`
-              <div class="indicator horizontal center layout">
-                <mwc-icon-button
-                  class="fg controls-running"
-                  icon="insert_drive_file"
-                ></mwc-icon-button>
-                ${rowData.item.filename}
-              </div>
-            `}
-      `,
-      root,
-    );
-  }
-
-  /**
    * Render permission view - r, w, d.
    *
    * @param {Element} root - the row details content DOM element
@@ -2439,51 +1710,6 @@ export default class BackendAiStorageList extends BackendAIPage {
   }
 
   /**
-   * Render created time.
-   *
-   * @param {Element} root - the row details content DOM element
-   * @param {Element} column - the column element that controls the state of the host element
-   * @param {Object} rowData - the object with the properties related with the rendered item
-   * */
-  createdTimeRenderer(root, column?, rowData?) {
-    render(
-      // language=HTML
-      html`
-        <div class="layout vertical">
-          <span>${this._humanReadableTime(rowData.item.ctime)}</span>
-        </div>
-      `,
-      root,
-    );
-  }
-
-  /**
-   * Render size by condition
-   *
-   * @param {Element} root - the row details content DOM element
-   * @param {Element} column - the column element that controls the state of the host element
-   * @param {Object} rowData - the object with the properties related with the rendered item
-   * */
-  sizeRenderer(root, column?, rowData?) {
-    render(
-      // language=HTML
-      html`
-        <div class="layout horizontal">
-          ${(rowData.item.type as string).toUpperCase() === 'DIRECTORY' &&
-          !this._isDirectorySizeVisible
-            ? html`
-                <span class="monospace">-</span>
-              `
-            : html`
-                <span>${rowData.item.size}</span>
-              `}
-        </div>
-      `,
-      root,
-    );
-  }
-
-  /**
    * Render type of user - person, group.
    *
    * @param {Element} root - the row details content DOM element
@@ -2506,59 +1732,6 @@ export default class BackendAiStorageList extends BackendAIPage {
       `,
       root,
     );
-  }
-
-  private async _getCurrentKeypairResourcePolicy() {
-    const accessKey = globalThis.backendaiclient._config.accessKey;
-    const res = await globalThis.backendaiclient.keypair.info(accessKey, [
-      'resource_policy',
-    ]);
-    return res.keypair.resource_policy;
-  }
-
-  async _getVolumeInformation() {
-    const vhostInfo = await globalThis.backendaiclient.vfolder.list_hosts();
-    this.volumeInfo = vhostInfo.volume_info || {};
-  }
-
-  async _getAllowedVFolderHostsByCurrentUserInfo() {
-    const [vhostInfo, currentKeypairResourcePolicy] = await Promise.all([
-      globalThis.backendaiclient.vfolder.list_hosts(),
-      this._getCurrentKeypairResourcePolicy(),
-    ]);
-    const currentDomain = globalThis.backendaiclient._config.domainName;
-    const currentGroupId = globalThis.backendaiclient.current_group_id();
-    const mergedData =
-      await globalThis.backendaiclient.storageproxy.getAllowedVFolderHostsByCurrentUserInfo(
-        currentDomain,
-        currentGroupId,
-        currentKeypairResourcePolicy,
-      );
-
-    const allowedPermissionForDomainsByVolume = JSON.parse(
-      mergedData?.domain?.allowed_vfolder_hosts || '{}',
-    );
-    const allowedPermissionForGroupsByVolume = JSON.parse(
-      mergedData?.group?.allowed_vfolder_hosts || '{}',
-    );
-    const allowedPermissionForResourcePolicyByVolume = JSON.parse(
-      mergedData?.keypair_resource_policy.allowed_vfolder_hosts || '{}',
-    );
-
-    const _mergeDedupe = (arr) => [...new Set([].concat(...arr))];
-    this._unionedAllowedPermissionByVolume = Object.assign(
-      {},
-      ...vhostInfo.allowed.map((volume) => {
-        return {
-          [volume]: _mergeDedupe([
-            allowedPermissionForDomainsByVolume[volume],
-            allowedPermissionForGroupsByVolume[volume],
-            allowedPermissionForResourcePolicyByVolume[volume],
-          ]),
-        };
-      }),
-    );
-    this.folderListGrid.clearCache();
   }
 
   _checkFolderSupportDirectoryBasedUsage(host: string) {
@@ -2675,55 +1848,6 @@ export default class BackendAiStorageList extends BackendAIPage {
     });
   }
 
-  _refreshFolderUI(e) {
-    if (
-      Object.prototype.hasOwnProperty.call(e.detail, 'mini-ui') &&
-      e.detail['mini-ui'] === true
-    ) {
-      this.folderExplorerDialog.classList.add('mini_ui');
-    } else {
-      this.folderExplorerDialog.classList.remove('mini_ui');
-    }
-  }
-
-  /**
-   * Check the images that supports filebrowser application
-   *
-   */
-  async _checkImageSupported() {
-    const fields = [
-      'name',
-      'tag',
-      'registry',
-      'digest',
-      'installed',
-      'labels { key value }',
-      'resource_limits { key min max }',
-    ];
-    const response = await globalThis.backendaiclient.image.list(
-      fields,
-      true,
-      true,
-    );
-    const images = response.images;
-    // Filter filebrowser supported images.
-    this.filebrowserSupportedImages = images.filter((image) =>
-      image.labels.find(
-        (label) =>
-          label.key === 'ai.backend.service-ports' &&
-          label.value.toLowerCase().includes('filebrowser'),
-      ),
-    );
-    // Filter service supported images.
-    this.systemRoleSupportedImages = images.filter((image) =>
-      image.labels.find(
-        (label) =>
-          label.key === 'ai.backend.role' &&
-          label.value.toLowerCase().includes('system'),
-      ),
-    );
-  }
-
   async _viewStateChanged(active) {
     await this.updateComplete;
     if (active === false) {
@@ -2745,19 +1869,11 @@ export default class BackendAiStorageList extends BackendAIPage {
           this.enableVfolderTrashBin =
             globalThis.backendaiclient.supports('vfolder-trash-bin');
           this.authenticated = true;
-          this._APIMajorVersion = globalThis.backendaiclient.APIMajorVersion;
-          this._maxFileUploadSize =
-            globalThis.backendaiclient._config.maxFileUploadSize;
           this.directoryBasedUsage =
             globalThis.backendaiclient._config.directoryBasedUsage &&
             !globalThis.backendaiclient.supports(
               'deprecated-max-quota-scope-in-keypair-resource-policy',
             );
-          this._isDirectorySizeVisible =
-            globalThis.backendaiclient._config.isDirectorySizeVisible;
-          this._getAllowedVFolderHostsByCurrentUserInfo();
-          this._checkImageSupported();
-          this._getVolumeInformation();
           this._refreshFolderList(false, 'viewStatechanged');
         },
         true,
@@ -2771,30 +1887,13 @@ export default class BackendAiStorageList extends BackendAIPage {
       this.enableVfolderTrashBin =
         globalThis.backendaiclient.supports('vfolder-trash-bin');
       this.authenticated = true;
-      this._APIMajorVersion = globalThis.backendaiclient.APIMajorVersion;
-      this._maxFileUploadSize =
-        globalThis.backendaiclient._config.maxFileUploadSize;
       this.directoryBasedUsage =
         globalThis.backendaiclient._config.directoryBasedUsage &&
         !globalThis.backendaiclient.supports(
           'deprecated-max-quota-scope-in-keypair-resource-policy',
         );
-      this._isDirectorySizeVisible =
-        globalThis.backendaiclient._config.isDirectorySizeVisible;
-      this._getAllowedVFolderHostsByCurrentUserInfo();
-      this._checkImageSupported();
-      this._getVolumeInformation();
       this._refreshFolderList(false, 'viewStatechanged');
     }
-  }
-
-  _folderExplorerDialog() {
-    this.openDialog('folder-explorer-dialog');
-  }
-
-  _mkdirDialog() {
-    this.mkdirNameInput.value = '';
-    this.openDialog('mkdir-dialog');
   }
 
   openDialog(id) {
@@ -3242,67 +2341,6 @@ export default class BackendAiStorageList extends BackendAIPage {
   }
 
   /**
-   * Validate file/subfolder name.
-   */
-  _validateExistingFileName() {
-    this.newFileNameInput.validityTransform = (newValue, nativeValidity) => {
-      if (!nativeValidity.valid) {
-        if (nativeValidity.valueMissing) {
-          this.newFileNameInput.validationMessage = _text(
-            'data.FileandFoldernameRequired',
-          );
-          return {
-            valid: nativeValidity.valid,
-            customError: !nativeValidity.valid,
-          };
-        } else {
-          this.newFileNameInput.validationMessage = _text(
-            'data.Allowslettersnumbersand-_dot',
-          );
-          return {
-            valid: nativeValidity.valid,
-            customError: !nativeValidity.valid,
-          };
-        }
-      } else {
-        const regex = /[`~!@#$%^&*()|+=?;:'",<>{}[\]\\/]/gi;
-        let isValid: boolean;
-        // compare old name and new name.
-        if (
-          this.newFileNameInput.value ===
-          (
-            this.renameFileDialog.querySelector(
-              '#old-file-name',
-            ) as HTMLDivElement
-          ).textContent
-        ) {
-          this.newFileNameInput.validationMessage = _text(
-            'data.EnterDifferentValue',
-          );
-          isValid = false;
-          return {
-            valid: isValid,
-            customError: !isValid,
-          };
-        } else {
-          isValid = true;
-        }
-        // custom validation for folder name using regex
-        isValid = !regex.test(this.newFileNameInput.value);
-        if (!isValid) {
-          this.newFileNameInput.validationMessage = _text(
-            'data.Allowslettersnumbersand-_dot',
-          );
-        }
-        return {
-          valid: isValid,
-          customError: !isValid,
-        };
-      }
-    };
-  }
-
-  /**
    * Validate folder name.
    *
    * @param {boolean} isModifying
@@ -3360,77 +2398,6 @@ export default class BackendAiStorageList extends BackendAIPage {
     };
   }
 
-  /* Folder Explorer*/
-  /**
-   * Clear the folder explorer.
-   *
-   * @param {string} path - explorer path
-   * @param {string} id - explorer id
-   * @param {boolean} dialog - whether open folder-explorer-dialog or not
-   * */
-  async _clearExplorer(
-    path = this.explorer.breadcrumb.join('/'),
-    id = this.explorer.id,
-    dialog = false,
-  ) {
-    const job = await globalThis.backendaiclient.vfolder.list_files(path, id);
-    this.fileListGrid.selectedItems = [];
-    if (this._APIMajorVersion < 6) {
-      this.explorer.files = JSON.parse(job.files);
-    } else {
-      // to support dedicated storage vendors such as FlashBlade
-      const fileInfo = JSON.parse(job.files);
-      fileInfo.forEach((info, cnt) => {
-        let ftype = 'FILE';
-        if (info.filename === job.items[cnt].name) {
-          // value.files and value.items have same order
-          ftype = job.items[cnt].type;
-        } else {
-          // In case the order is mixed
-          for (let i = 0; i < job.items.length; i++) {
-            if (info.filename === job.items[i].name) {
-              ftype = job.items[i].type;
-              break;
-            }
-          }
-        }
-        info.type = ftype;
-      });
-      this.explorer.files = fileInfo;
-    }
-    this.explorerFiles = this.explorer.files;
-    if (dialog) {
-      if (
-        this.filebrowserSupportedImages.length === 0 ||
-        this.systemRoleSupportedImages.length === 0
-      ) {
-        await this._checkImageSupported();
-      }
-      this._toggleFilebrowserButton();
-      this._toggleSSHSessionButton();
-      this.openDialog('folder-explorer-dialog');
-    }
-  }
-
-  /**
-   * toggle filebrowser button in Vfolder explorer dialog
-   */
-  _toggleFilebrowserButton() {
-    const isfilebrowserSupported =
-      this.filebrowserSupportedImages.length > 0 && this._isResourceEnough()
-        ? true
-        : false;
-    const filebrowserIcon = this.shadowRoot?.querySelector('#filebrowser-img');
-    const filebrowserBtn = this.shadowRoot?.querySelector(
-      '#filebrowser-btn',
-    ) as Button;
-    if (filebrowserIcon && filebrowserBtn) {
-      filebrowserBtn.disabled = !isfilebrowserSupported;
-      const filterClass = isfilebrowserSupported ? '' : 'apply-grayscale';
-      filebrowserIcon.setAttribute('class', filterClass);
-    }
-  }
-
   triggerOpenFilebrowserToReact(rowData) {
     const queryParams = new URLSearchParams(window.location.search);
     queryParams.set('folder', rowData.item.id);
@@ -3442,1001 +2409,6 @@ export default class BackendAiStorageList extends BackendAIPage {
         },
       }),
     );
-  }
-
-  triggerCloseFilebrowserToReact() {
-    const queryParams = new URLSearchParams(window.location.search);
-    queryParams.delete('folder');
-    document.dispatchEvent(
-      new CustomEvent('react-navigate', {
-        detail: {
-          pathname: window.location.pathname,
-          search: queryParams.toString(),
-        },
-      }),
-    );
-  }
-
-  /**
-   * Set up the explorer of the folder and call the _clearExplorer() function.
-   *
-   * @param {Event} e - click the folder_open icon button
-   * @param {boolean} isWritable - check whether write operation is allowed or not
-   * */
-  _folderExplorer(rowData) {
-    this.vhost = rowData.item.host;
-    const folderName = rowData.item.name;
-    const isWritable =
-      this._hasPermission(rowData.item, 'w') ||
-      rowData.item.is_owner ||
-      (rowData.item.type === 'group' && this.is_admin);
-    const explorer = {
-      id: folderName,
-      uuid: rowData.item.id,
-      breadcrumb: ['.'],
-    };
-
-    /**
-     * NOTICE: If it's admin user and the folder type is group, It will have write permission.
-     */
-    this.isWritable = isWritable;
-    this.explorer = explorer;
-    this._clearExplorer(explorer.breadcrumb.join('/'), explorer.id, true);
-  }
-
-  /**
-   * Enqueue the folder and call the _clearExplorer() function.
-   *
-   * @param {Event} e - click the folder_open icon button
-   * */
-  _enqueueFolder(e) {
-    const button = e.target;
-
-    // disable button to avoid executing extra onclick event
-    button.setAttribute('disabled', 'true');
-    const fn = e.target.getAttribute('name');
-    this.explorer.breadcrumb.push(fn);
-
-    // enable button only if the operation is done.
-    this._clearExplorer().then((res) => {
-      button.removeAttribute('disabled');
-    });
-  }
-
-  _gotoFolder(e) {
-    const dest = e.target.getAttribute('dest');
-    let tempBreadcrumb = this.explorer.breadcrumb;
-    const index = tempBreadcrumb.indexOf(dest);
-
-    if (index === -1) {
-      return;
-    }
-
-    tempBreadcrumb = tempBreadcrumb.slice(0, index + 1);
-
-    this.explorer.breadcrumb = tempBreadcrumb;
-    this._clearExplorer(tempBreadcrumb.join('/'), this.explorer.id, false);
-  }
-
-  _mkdir(e) {
-    const newfolder = this.mkdirNameInput.value;
-    const explorer = this.explorer;
-    this.mkdirNameInput.reportValidity();
-    if (this.mkdirNameInput.checkValidity()) {
-      const job = globalThis.backendaiclient.vfolder
-        .mkdir([...explorer.breadcrumb, newfolder].join('/'), explorer.id)
-        .catch((err) => {
-          if (err & err.message) {
-            this.notification.text = PainKiller.relieve(err.title);
-            this.notification.detail = err.message;
-            this.notification.show(true, err);
-          } else if (err && err.title) {
-            this.notification.text = PainKiller.relieve(err.title);
-            this.notification.show(true, err);
-          }
-        });
-      job.then((res) => {
-        this.closeDialog('mkdir-dialog');
-        this._clearExplorer();
-      });
-    } else {
-      return;
-    }
-  }
-
-  _isDir(file) {
-    if (this._APIMajorVersion < 6) {
-      return file.mode.startsWith('d');
-    } else {
-      // For some vendor-specific storage APIs, we cannot discern file and
-      // directory by just looking at the first letter of file mode. For
-      // example, FlashBlade API returns file's mode as a pure number.
-      // So, we have to rely on the explicit file's type property to support
-      // vendor-specific APIs.
-      return file.type === 'DIRECTORY';
-    }
-  }
-
-  /* File upload and download */
-  /**
-   * Add eventListener to the dropzone - dragleave, dragover, drop.
-   * */
-  _addEventListenerDropZone() {
-    const dndZonePlaceholderEl = this.shadowRoot?.querySelector(
-      '#dropzone',
-    ) as HTMLDivElement;
-    dndZonePlaceholderEl.addEventListener('dragleave', () => {
-      dndZonePlaceholderEl.style.display = 'none';
-    });
-
-    // TODO specify custom event type
-    this.folderExplorerDialog.addEventListener('dragover', (e: any) => {
-      e.stopPropagation();
-      e.preventDefault();
-      if (this.isWritable) {
-        e.dataTransfer.dropEffect = 'copy';
-        dndZonePlaceholderEl.style.display = 'flex';
-        return false;
-      } else {
-        return true;
-      }
-    });
-
-    // TODO specify custom event type
-    this.folderExplorerDialog.addEventListener('drop', (e: any) => {
-      let isNotificationDisplayed = false;
-      e.stopPropagation();
-      e.preventDefault();
-      dndZonePlaceholderEl.style.display = 'none';
-      if (this.isWritable) {
-        for (let i = 0; i < e.dataTransfer.files.length; i++) {
-          if (e.dataTransfer.items[i].webkitGetAsEntry().isFile) {
-            const file = e.dataTransfer.files[i];
-            /* Drag & Drop file upload size limits to configuration */
-            if (
-              this._maxFileUploadSize > 0 &&
-              file.size > this._maxFileUploadSize
-            ) {
-              this.notification.text =
-                _text('data.explorer.FileUploadSizeLimit') +
-                ` (${globalThis.backendaiutils._humanReadableFileSize(
-                  this._maxFileUploadSize,
-                )})`;
-              this.notification.show();
-              return;
-            } else {
-              const reUploadFile = this.explorerFiles.find(
-                (elem: any) => elem.filename === file.name,
-              );
-              if (reUploadFile) {
-                // plain javascript modal to confirm whether proceed to overwrite operation or not
-                /*
-                 *  TODO: replace confirm operation with customized dialog
-                 */
-                const confirmed = window.confirm(
-                  `${_text('data.explorer.FileAlreadyExists')}\n${
-                    file.name
-                  }\n${_text('data.explorer.DoYouWantToOverwrite')}`,
-                );
-                if (confirmed) {
-                  file.progress = 0;
-                  file.caption = '';
-                  file.error = false;
-                  file.complete = false;
-                  this.uploadFiles.push(file);
-                }
-              } else {
-                file.progress = 0;
-                file.caption = '';
-                file.error = false;
-                file.complete = false;
-                this.uploadFiles.push(file);
-              }
-            }
-          } else {
-            // let item = e.dataTransfer.items[i].webkitGetAsEntry();
-            // console.log(item.webkitRelativePath);
-            // this._executeFileBrowser();
-            // show snackbar to filebrowser only once
-            if (!isNotificationDisplayed) {
-              if (this.filebrowserSupportedImages.length > 0) {
-                this.notification.text = _text(
-                  'data.explorer.ClickFilebrowserButton',
-                );
-                this.notification.show();
-              } else {
-                this.notification.text = _text(
-                  'data.explorer.NoImagesSupportingFileBrowser',
-                );
-                this.notification.show();
-              }
-            }
-            isNotificationDisplayed = true;
-          }
-        }
-
-        for (let i = 0; i < this.uploadFiles.length; i++) {
-          this.fileUpload(this.uploadFiles[i]);
-          this._clearExplorer();
-        }
-      } else {
-        this.notification.text = _text(
-          'data.explorer.WritePermissionRequiredInUploadFiles',
-        );
-        this.notification.show();
-      }
-    });
-  }
-
-  /**
-   * Create MouseEvents when cloud_upload button is clicked.
-   *
-   * @param {Event} e - click the cloud_upload button.
-   * */
-  _uploadBtnClick(e) {
-    const isFolder = e.target.id === 'add-folder-btn';
-    const elem = isFolder
-      ? (this.shadowRoot?.querySelector('#folderInput') as HTMLInputElement)
-      : (this.shadowRoot?.querySelector('#fileInput') as HTMLInputElement);
-    if (elem && document.createEvent) {
-      // sanity check
-      const evt = document.createEvent('MouseEvents');
-      evt.initEvent('click', true, false);
-      elem.dispatchEvent(evt);
-    }
-  }
-
-  getFolderName(file: File) {
-    const filePath = file.webkitRelativePath || file.name;
-    return filePath.split('/')?.[0];
-  }
-
-  /**
-   * If file is added, call the fileUpload() function and initialize fileInput string
-   *
-   * @param {Event} e - add file to the input element
-   * */
-  _uploadInputChange(e) {
-    const length = e.target.files.length;
-    const isFolderUpload = e.target.id === 'folderInput';
-    const inputElement = isFolderUpload
-      ? (this.shadowRoot?.querySelector('#folderInput') as HTMLInputElement)
-      : (this.shadowRoot?.querySelector('#fileInput') as HTMLInputElement);
-    let isEmptyFileIncluded = false;
-    let reUploadFolderConfirmed = false;
-    // plain javascript modal to confirm whether proceed to overwrite "folder" operation or not
-    /*
-     *  TODO: replace confirm operation with customized dialog
-     */
-    if (e.target.files.length > 0 && isFolderUpload) {
-      const f = e.target.files[0];
-      const reUploadFolder = this.explorerFiles.find(
-        (elem: any) => elem.filename === this.getFolderName(f),
-      );
-      if (reUploadFolder) {
-        reUploadFolderConfirmed = window.confirm(
-          `${_text('data.explorer.FolderAlreadyExists')}\n${this.getFolderName(
-            f,
-          )}\n${_text('data.explorer.DoYouWantToOverwrite')}`,
-        );
-        if (!reUploadFolderConfirmed) {
-          inputElement.value = '';
-          return;
-        }
-      }
-    }
-    for (let i = 0; i < length; i++) {
-      const file = e.target.files[i];
-
-      let text = '';
-      const possible =
-        'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-      for (let i = 0; i < 5; i++) {
-        text += possible.charAt(Math.floor(Math.random() * possible.length));
-      }
-      /* File upload size limits to configuration */
-      if (this._maxFileUploadSize > 0 && file.size > this._maxFileUploadSize) {
-        this.notification.text =
-          _text('data.explorer.FileUploadSizeLimit') +
-          ` (${globalThis.backendaiutils._humanReadableFileSize(
-            this._maxFileUploadSize,
-          )})`;
-        this.notification.show();
-        return;
-      } else if (file.size === 0) {
-        // skip the empty file upload
-        isEmptyFileIncluded = true;
-        continue;
-      } else {
-        const reUploadFile = this.explorerFiles.find(
-          (elem: any) => elem.filename === file.name,
-        );
-        if (reUploadFile && !reUploadFolderConfirmed) {
-          // plain javascript modal to confirm whether proceed to overwrite "file" operation or not
-          // if the user already confirms to overwrite the "folder", the modal doesn't appear.
-          /*
-           *  TODO: replace confirm operation with customized dialog
-           */
-          const confirmed = window.confirm(
-            `${_text('data.explorer.FileAlreadyExists')}\n${file.name}\n${_text(
-              'data.explorer.DoYouWantToOverwrite',
-            )}`,
-          );
-          if (confirmed) {
-            file.id = text;
-            file.progress = 0;
-            file.caption = '';
-            file.error = false;
-            file.complete = false;
-            this.uploadFiles.push(file);
-          }
-        } else {
-          file.id = text;
-          file.progress = 0;
-          file.caption = '';
-          file.error = false;
-          file.complete = false;
-          this.uploadFiles.push(file);
-        }
-      }
-    }
-    for (let i = 0; i < this.uploadFiles.length; i++) {
-      this.fileUpload(this.uploadFiles[i]);
-    }
-    if (isEmptyFileIncluded || isFolderUpload) {
-      this.notification.text = _text(
-        'data.explorer.EmptyFilesAndFoldersAreNotUploaded',
-      );
-      this.notification.show();
-    }
-    inputElement.value = '';
-  }
-
-  /**
-   * Running file upload queue to upload files.
-   *
-   * @param {null | string} session - upload session
-   * */
-  runFileUploadQueue(session = null) {
-    if (session !== null) {
-      (this.fileUploadQueue as any).push(session);
-    }
-    let queuedSession;
-    for (
-      let i = this.fileUploadCount;
-      i < this.concurrentFileUploadLimit;
-      i++
-    ) {
-      if (this.fileUploadQueue.length > 0) {
-        queuedSession = this.fileUploadQueue.shift();
-        this.fileUploadCount = this.fileUploadCount + 1;
-        queuedSession.start();
-      }
-    }
-  }
-
-  /**
-   * Upload the file.
-   *
-   * @param {Object} fileObj - file object
-   * */
-  fileUpload(fileObj) {
-    this._uploadFlag = true;
-    this.uploadFilesExist = this.uploadFiles.length > 0;
-    const path = this.explorer.breadcrumb
-      .concat(fileObj.webkitRelativePath || fileObj.name)
-      .join('/');
-    const job = globalThis.backendaiclient.vfolder.create_upload_session(
-      path,
-      fileObj,
-      this.explorer.id,
-    );
-    job.then((url) => {
-      const start_date = new Date().getTime();
-      const upload = new tus.Upload(fileObj, {
-        endpoint: url,
-        retryDelays: [0, 3000, 5000, 10000, 20000],
-        uploadUrl: url,
-        chunkSize: 15728640, // 15MB
-        metadata: {
-          filename: path,
-          filetype: fileObj.type,
-        },
-        onError: (error) => {
-          console.log('Failed because: ' + error);
-          this.currentUploadFile =
-            this.uploadFiles[this.uploadFiles.indexOf(fileObj)];
-          this.fileUploadCount = this.fileUploadCount - 1;
-          this.runFileUploadQueue();
-        },
-        onProgress: (bytesUploaded, bytesTotal) => {
-          this.currentUploadFile =
-            this.uploadFiles[this.uploadFiles.indexOf(fileObj)];
-          if (!this._uploadFlag) {
-            upload.abort();
-            this.uploadFiles[this.uploadFiles.indexOf(fileObj)].caption =
-              `Canceling...`;
-            this.uploadFiles = this.uploadFiles.slice();
-            setTimeout(() => {
-              this.uploadFiles = [];
-              this.uploadFilesExist = false;
-              this.fileUploadCount = this.fileUploadCount - 1;
-            }, 1000);
-            return;
-          }
-
-          const now = new Date().getTime();
-          const speed: string =
-            (
-              bytesUploaded /
-              (1024 * 1024) /
-              ((now - start_date) / 1000)
-            ).toFixed(1) + 'MB/s';
-          const estimated_seconds = Math.floor(
-            (bytesTotal - bytesUploaded) /
-              ((bytesUploaded / (now - start_date)) * 1000),
-          );
-          let estimated_time_left = _text('data.explorer.LessThan10Sec');
-          if (estimated_seconds >= 86400) {
-            estimated_time_left = _text('data.explorer.MoreThanADay');
-          } else if (estimated_seconds > 10) {
-            const hour = Math.floor(estimated_seconds / 3600);
-            const min = Math.floor((estimated_seconds % 3600) / 60);
-            const sec = estimated_seconds % 60;
-            estimated_time_left = `${hour}:${min}:${sec}`;
-          }
-          const percentage = ((bytesUploaded / bytesTotal) * 100).toFixed(1);
-          this.uploadFiles[this.uploadFiles.indexOf(fileObj)].progress =
-            bytesUploaded / bytesTotal;
-          this.uploadFiles[this.uploadFiles.indexOf(fileObj)].caption =
-            `${percentage}% / Time left : ${estimated_time_left} / Speed : ${speed}`;
-          this.uploadFiles = this.uploadFiles.slice();
-        },
-        onSuccess: () => {
-          this._clearExplorer();
-          this.currentUploadFile =
-            this.uploadFiles[this.uploadFiles.indexOf(fileObj)];
-          this.uploadFiles[this.uploadFiles.indexOf(fileObj)].complete = true;
-          this.uploadFiles = this.uploadFiles.slice();
-          setTimeout(() => {
-            this.uploadFiles.splice(this.uploadFiles.indexOf(fileObj), 1);
-            this.uploadFilesExist = this.uploadFiles.length > 0 ? true : false;
-            this.uploadFiles = this.uploadFiles.slice();
-            this.fileUploadCount = this.fileUploadCount - 1;
-            this.runFileUploadQueue();
-          }, 1000);
-        },
-      });
-      this.runFileUploadQueue(upload);
-    });
-  }
-
-  /**
-   * Cancel upload files.
-   *
-   * */
-  _cancelUpload() {
-    this._uploadFlag = false;
-  }
-
-  /**
-   * Download the file.
-   *
-   * @param {Event} e - click the cloud_download icon button
-   * @param {boolean} archive - whether archive or not
-   * */
-  _downloadFile(e, archive = false) {
-    if (!this._isDownloadable(this.vhost)) {
-      this.notification.text = _text('data.explorer.DownloadNotAllowed');
-      this.notification.show();
-      return;
-    }
-    const fn = e.target.getAttribute('filename');
-    const path = this.explorer.breadcrumb.concat(fn).join('/');
-    const job = globalThis.backendaiclient.vfolder.request_download_token(
-      path,
-      this.explorer.id,
-      archive,
-    );
-    job.then((res) => {
-      const token = res.token;
-      let url;
-      if (this._APIMajorVersion < 6) {
-        url =
-          globalThis.backendaiclient.vfolder.get_download_url_with_token(token);
-      } else {
-        url = `${res.url}?token=${res.token}&archive=${archive}`;
-      }
-      if (globalThis.iOSSafari) {
-        this.downloadURL = url;
-        this.downloadFileDialog.show();
-        URL.revokeObjectURL(url);
-      } else {
-        const a = document.createElement('a');
-        a.style.display = 'none';
-        a.addEventListener('click', function (e) {
-          e.stopPropagation();
-        });
-        a.href = url;
-        a.download = fn;
-        document.body.appendChild(a); // we need to append the element to the dom -> otherwise it will not work in firefox
-        a.click();
-        // a.remove();  //afterwards we remove the element again
-        document.body.removeChild(a);
-        URL.revokeObjectURL(url);
-      }
-    });
-  }
-
-  /**
-   * Select filename without extension
-   *
-   */
-  _compareFileExtension() {
-    const newFilename = this.newFileNameInput.value;
-    const oldFilename =
-      (this.renameFileDialog.querySelector('#old-file-name') as HTMLDivElement)
-        .textContent ?? '';
-    const regex = /\.([0-9a-z]+)$/i;
-    const newFileExtension = newFilename.match(regex);
-    const oldFileExtension = oldFilename.match(regex);
-    if (newFilename.includes('.') && newFileExtension) {
-      this.newFileExtension = newFileExtension[1].toLowerCase();
-    } else {
-      this.newFileExtension = '';
-    }
-    if (oldFilename.includes('.') && oldFileExtension) {
-      this.oldFileExtension = oldFileExtension[1].toLowerCase();
-    } else {
-      this.oldFileExtension = '';
-    }
-
-    if (newFilename) {
-      if (this.newFileExtension !== this.oldFileExtension) {
-        this.fileExtensionChangeDialog.show();
-      } else if (this.oldFileExtension) {
-        this._keepFileExtension();
-      } else {
-        this._renameFile();
-      }
-    } else {
-      this._renameFile();
-    }
-  }
-
-  /**
-   * Keep the file extension whether the file extension is explicit or not.
-   *
-   */
-  _keepFileExtension() {
-    let newFilename = this.newFileNameInput.value;
-    if (this.newFileExtension) {
-      newFilename = newFilename.replace(
-        new RegExp(this.newFileExtension + '$'),
-        this.oldFileExtension,
-      );
-    } else {
-      newFilename = newFilename + '.' + this.oldFileExtension;
-    }
-    this.newFileNameInput.value = newFilename;
-    this._renameFile();
-  }
-
-  /* Execute Filebrowser by launching session with mimimum resources
-   *
-   */
-  _executeFileBrowser() {
-    if (this._isResourceEnough()) {
-      if (this.filebrowserSupportedImages.length > 0) {
-        const isNotificationVisible = localStorage.getItem(
-          'backendaiwebui.filebrowserNotification',
-        );
-        if (
-          (isNotificationVisible == null || isNotificationVisible === 'true') &&
-          !this.isWritable
-        ) {
-          this.fileBrowserNotificationDialog.show();
-        }
-        this._launchFileBrowserSession();
-        this._toggleFilebrowserButton();
-      } else {
-        this.notification.text = _text(
-          'data.explorer.NoImagesSupportingFileBrowser',
-        );
-        this.notification.show();
-      }
-    } else {
-      this.notification.text = _text(
-        'data.explorer.NotEnoughResourceForFileBrowserSession',
-      );
-      this.notification.show();
-    }
-  }
-
-  /**
-   * Toggle notification of filebrowser execution on read-only folder
-   *
-   * @param {any} e
-   */
-  _toggleShowFilebrowserNotification(e) {
-    const checkbox = e.target;
-    if (checkbox) {
-      const isHidden = (!checkbox.checked).toString();
-      localStorage.setItem('backendaiwebui.filebrowserNotification', isHidden);
-    }
-  }
-
-  /**
-   * Open the session launcher dialog to execute filebrowser app.
-   *
-   */
-  async _launchFileBrowserSession() {
-    let appOptions;
-    const imageResource: Record<string, unknown> = {};
-    // monkeypatch for filebrowser applied environment
-    // const environment = 'cr.backend.ai/testing/filebrowser:21.01-ubuntu20.04';
-    const images = this.filebrowserSupportedImages.filter(
-      (image: any) =>
-        image['name'].toLowerCase().includes('filebrowser') &&
-        image['installed'],
-    );
-
-    // select one image to launch filebrowser supported session
-    const preferredImage = images[0];
-    const environment =
-      preferredImage['registry'] +
-      '/' +
-      preferredImage['name'] +
-      ':' +
-      preferredImage['tag'];
-
-    // add current folder
-    imageResource['mounts'] = [this.explorer.id];
-    imageResource['cpu'] = 1;
-    imageResource['mem'] = this.minimumResource.mem + 'g';
-    imageResource['domain'] = globalThis.backendaiclient._config.domainName;
-    imageResource['group_name'] = globalThis.backendaiclient.current_group;
-    const indicator = await this.indicator.start('indeterminate');
-
-    return globalThis.backendaiclient
-      .get_resource_slots()
-      .then((response) => {
-        indicator.set(20, _text('data.explorer.ExecutingFileBrowser'));
-        return globalThis.backendaiclient.createIfNotExists(
-          environment,
-          null,
-          imageResource,
-          10000,
-          undefined,
-        );
-      })
-      .then(async (res) => {
-        const service_info = res.servicePorts;
-        appOptions = {
-          'session-uuid': res.sessionId,
-          'session-name': res.sessionName,
-          'access-key': '',
-          runtime: 'filebrowser',
-          arguments: { '--root': '/home/work/' + this.explorer.id },
-        };
-        // only launch filebrowser app when it has valid service ports
-        if (
-          service_info.length > 0 &&
-          service_info.filter((el) => el.name === 'filebrowser').length > 0
-        ) {
-          globalThis.appLauncher.showLauncher(appOptions);
-        }
-        if (this.folderExplorerDialog.open) {
-          this.closeDialog('folder-explorer-dialog');
-        }
-        indicator.end(1000);
-      })
-      .catch((err) => {
-        this.notification.text = PainKiller.relieve(err.title);
-        this.notification.detail = err.message;
-        this.notification.show(true, err);
-        indicator.end(100);
-      });
-  }
-
-  _executeSSHProxyAgent() {
-    if (this.volumeInfo[this.vhost]?.sftp_scaling_groups?.length > 0) {
-      if (this.systemRoleSupportedImages.length > 0) {
-        this._launchSystemRoleSSHSession();
-        this._toggleSSHSessionButton();
-      } else {
-        this.notification.text = _text(
-          'data.explorer.NoImagesSupportingSystemSession',
-        );
-        this.notification.show();
-      }
-    } else {
-      this.notification.text = _text('data.explorer.SFTPSessionNotAvailable');
-      this.notification.show();
-    }
-  }
-
-  /**
-   * Launch system role sftp-uploader image and open the dialog that includes the ssh link.
-   */
-  async _launchSystemRoleSSHSession() {
-    const imageResource: Record<string, unknown> = {};
-    const configSSHImage = globalThis.backendaiclient._config.systemSSHImage;
-    const images = this.systemRoleSupportedImages.filter(
-      (image: any) => image['installed'],
-    );
-    // TODO: use lablup/openssh-server image
-    // select one image to launch system role supported session
-    const preferredImage = images[0];
-    const environment =
-      configSSHImage !== ''
-        ? configSSHImage
-        : preferredImage['registry'] +
-          '/' +
-          preferredImage['name'] +
-          ':' +
-          preferredImage['tag'];
-
-    // add current folder
-    imageResource['mounts'] = [this.explorer.id];
-    imageResource['cpu'] = 1;
-    imageResource['mem'] = '256m';
-    imageResource['domain'] = globalThis.backendaiclient._config.domainName;
-    imageResource['scaling_group'] =
-      this.volumeInfo[this.vhost]?.sftp_scaling_groups[0];
-    imageResource['group_name'] = globalThis.backendaiclient.current_group;
-    const indicator = await this.indicator.start('indeterminate');
-    return (async () => {
-      try {
-        await globalThis.backendaiclient.get_resource_slots();
-        indicator.set(50, _text('data.explorer.StartingSSH/SFTPSession'));
-        const sessionResponse =
-          await globalThis.backendaiclient.createIfNotExists(
-            environment,
-            `sftp-${this.explorer.uuid}`,
-            imageResource,
-            15000,
-            undefined,
-          );
-        if (sessionResponse.status === 'CANCELLED') {
-          // Max # of upload sessions exceeded for this used
-          this.notification.text = PainKiller.relieve(
-            _text('data.explorer.NumberOfSFTPSessionsExceededTitle'),
-          );
-          this.notification.detail = _text(
-            'data.explorer.NumberOfSFTPSessionsExceededBody',
-          );
-          this.notification.show(true, {
-            title: _text('data.explorer.NumberOfSFTPSessionsExceededTitle'),
-            message: _text('data.explorer.NumberOfSFTPSessionsExceededBody'),
-          });
-          indicator.end(100);
-          return;
-        }
-        const directAccessInfo =
-          await globalThis.backendaiclient.get_direct_access_info(
-            sessionResponse.sessionId,
-          );
-        const host = directAccessInfo.public_host.replace(/^https?:\/\//, '');
-        const port = directAccessInfo.sshd_ports;
-        const event = new CustomEvent('read-ssh-key-and-launch-ssh-dialog', {
-          detail: {
-            sessionUuid: sessionResponse.sessionId,
-            host: host,
-            port: port,
-            mounted: this.explorer.id,
-          },
-        });
-        document.dispatchEvent(event);
-        indicator.end(100);
-      } catch (err) {
-        this.notification.text = PainKiller.relieve(err.title);
-        this.notification.detail = err.message;
-        this.notification.show(true, err);
-        indicator.end(100);
-      }
-    })();
-  }
-
-  _toggleSSHSessionButton() {
-    const isSystemRoleSupported = this.systemRoleSupportedImages.length > 0;
-    const sshImageIcon = this.shadowRoot?.querySelector('#ssh-img');
-    const sshImageBtn = this.shadowRoot?.querySelector('#ssh-btn') as Button;
-    if (sshImageIcon && sshImageBtn) {
-      sshImageBtn.disabled = !isSystemRoleSupported;
-      const filterClass = isSystemRoleSupported ? '' : 'apply-grayscale';
-      sshImageIcon.setAttribute('class', filterClass);
-    }
-  }
-
-  /**
-   * Open the renameFileDialog to rename the file.
-   *
-   * @param {Event} e - click the edit icon button
-   * @param {Boolean} is_dir - True when file is directory type
-   * */
-  _openRenameFileDialog(e, is_dir = false) {
-    const fn = e.target.getAttribute('filename');
-    (
-      this.renameFileDialog.querySelector('#old-file-name') as HTMLDivElement
-    ).textContent = fn;
-    this.newFileNameInput.value = fn;
-    // TODO define extended type for custom property
-    this.renameFileDialog.filename = fn;
-    this.renameFileDialog.show();
-    this.is_dir = is_dir;
-
-    this.newFileNameInput.addEventListener('focus', (e) => {
-      const endOfExtensionLength = fn.replace(/\.([0-9a-z]+)$/i, '').length;
-      this.newFileNameInput.setSelectionRange(0, endOfExtensionLength);
-    });
-    this.newFileNameInput.focus();
-  }
-
-  /**
-   * Rename the file.
-   *
-   * */
-  _renameFile() {
-    // TODO define extended type for custom property
-    const fn = this.renameFileDialog.filename;
-    const path = this.explorer.breadcrumb.concat(fn).join('/');
-    const newName = this.newFileNameInput.value;
-    this.fileExtensionChangeDialog.hide();
-    this.newFileNameInput.reportValidity();
-    if (this.newFileNameInput.checkValidity()) {
-      if (fn === newName) {
-        this.newFileNameInput.focus();
-        this.notification.text = _text('data.folders.SameFileName');
-        this.notification.show();
-        return;
-      }
-
-      const job = globalThis.backendaiclient.vfolder.rename_file(
-        path,
-        newName,
-        this.explorer.id,
-        this.is_dir,
-      );
-      job
-        .then((res) => {
-          this.notification.text = _text('data.folders.FileRenamed');
-          this.notification.show();
-          this._clearExplorer();
-          this.renameFileDialog.hide();
-        })
-        .catch((err) => {
-          console.error(err);
-          if (err && err.message) {
-            this.notification.text = err.title;
-            this.notification.detail = err.message;
-            this.notification.show(true, err);
-          }
-        });
-    } else {
-      return;
-    }
-  }
-
-  /**
-   * Open delete file dialog to delete file.
-   *
-   * @param {Event} e - click the delete-btn
-   * */
-  _openDeleteFileDialog(e) {
-    const fn = e.target.getAttribute('filename');
-    // TODO define extended type for custom properties
-    this.deleteFileDialog.filename = fn;
-    this.deleteFileDialog.files = [];
-    this.deleteFileDialog.show();
-  }
-
-  /**
-   * Open the deleteFileDialog of selected files.
-   *
-   * @param {Event} e - click the delete button
-   * */
-  _openDeleteMultipleFileDialog(e?) {
-    // TODO define extended type for custom property
-    this.deleteFileDialog.files = this.fileListGrid.selectedItems;
-    this.deleteFileDialog.filename = '';
-    this.deleteFileDialog.show();
-  }
-
-  /**
-   * If the user presses the Delete File button, and the Okay button on the double check dialogue, delete the file.
-   *
-   * @param {Event} e - click the Okay button
-   * */
-
-  _deleteFileWithCheck(e) {
-    // TODO define extended type for custom property
-    const files = this.deleteFileDialog.files;
-    if (files.length > 0) {
-      const filenames: string[] = [];
-      files.forEach((file) => {
-        const filename = this.explorer.breadcrumb
-          .concat(file.filename)
-          .join('/');
-        filenames.push(filename);
-      });
-      const job = globalThis.backendaiclient.vfolder.delete_files(
-        filenames,
-        true,
-        this.explorer.id,
-      );
-      job.then((res) => {
-        this.notification.text =
-          files.length == 1
-            ? _text('data.folders.FileDeleted')
-            : _text('data.folders.MultipleFilesDeleted');
-        this.notification.show();
-        this._clearExplorer();
-        this.deleteFileDialog.hide();
-      });
-    } else {
-      // TODO define extended type for custom property
-      if (this.deleteFileDialog.filename != '') {
-        const path = this.explorer.breadcrumb
-          .concat(this.deleteFileDialog.filename)
-          .join('/');
-        const job = globalThis.backendaiclient.vfolder.delete_files(
-          [path],
-          true,
-          this.explorer.id,
-        );
-        job.then((res) => {
-          this.notification.text = _text('data.folders.FileDeleted');
-          this.notification.show();
-          this._clearExplorer();
-          this.deleteFileDialog.hide();
-        });
-      }
-    }
-  }
-
-  /**
-   * Delete a file.
-   *
-   * @param {HTMLElement} e - file list component that contains filename attribute
-   * */
-  _deleteFile(e) {
-    const fn = e.target.getAttribute('filename');
-    const path = this.explorer.breadcrumb.concat(fn).join('/');
-    const job = globalThis.backendaiclient.vfolder.delete_files(
-      [path],
-      true,
-      this.explorer.id,
-    );
-    job.then((res) => {
-      this.notification.text = _text('data.folders.FileDeleted');
-      this.notification.show();
-      this._clearExplorer();
-    });
-  }
-
-  /**
-   * Returns whether resource is enough to launch session for executing filebrowser app or not.
-   * @return {boolean} - true when resource is enough, false when resource is not enough to create a session.
-   */
-  _isResourceEnough() {
-    // update current resources statistics.
-    const event = new CustomEvent('backend-ai-calculate-current-resource');
-    document.dispatchEvent(event);
-    const currentResource = globalThis.backendaioptions.get('current-resource');
-    if (currentResource) {
-      currentResource.cpu =
-        typeof currentResource.cpu === 'string'
-          ? parseInt(currentResource['cpu'])
-          : currentResource['cpu'];
-      if (
-        currentResource.cpu >= this.minimumResource.cpu &&
-        currentResource.mem >= this.minimumResource.mem
-      ) {
-        return true;
-      }
-    }
-    return false;
   }
 
   /**
@@ -4451,19 +2423,6 @@ export default class BackendAiStorageList extends BackendAIPage {
     const hours = date.getHours();
     date.setHours(hours - offset);
     return date.toUTCString();
-  }
-
-  /**
-   * Return whether file is downloadable.
-   * NOTE: For now, It's handled by storage host, not file itself.
-   *
-   * @param {Object} host
-   * @return {boolean} true
-   * */
-  _isDownloadable(host) {
-    return (this._unionedAllowedPermissionByVolume[host] ?? []).includes(
-      'download-file',
-    );
   }
 
   /**
@@ -4618,45 +2577,6 @@ export default class BackendAiStorageList extends BackendAIPage {
         }
         this.notification.show();
       });
-  }
-
-  /**
-   * Validate path name
-   * */
-  _validatePathName() {
-    this.mkdirNameInput.validityTransform = (newValue, nativeValidity) => {
-      if (!nativeValidity.valid) {
-        if (nativeValidity.valueMissing) {
-          this.mkdirNameInput.validationMessage = _text(
-            'data.explorer.ValueRequired',
-          );
-          return {
-            valid: nativeValidity.valid,
-            customError: !nativeValidity.valid,
-          };
-        } else {
-          return {
-            valid: nativeValidity.valid,
-            customError: !nativeValidity.valid,
-          };
-        }
-      } else {
-        // custom validation for path name using regex
-        const regex =
-          /^([^`~!@#$%^&*()|+=?;:'",<>{}[\]\r\n/]{1,})+(\/[^`~!@#$%^&*()|+=?;:'",<>{}[\]\r\n/]{1,})*([/,\\]{0,1})$/gm;
-        let isValid = regex.test(this.mkdirNameInput.value);
-        if (!isValid || this.mkdirNameInput.value === './') {
-          this.mkdirNameInput.validationMessage = _text(
-            'data.explorer.ValueShouldBeStarted',
-          );
-          isValid = false;
-        }
-        return {
-          valid: isValid,
-          customError: !isValid,
-        };
-      }
-    };
   }
 
   /**
