+++ src/lib/backend.ai-client-esm.ts
@@ -2434,8 +2434,8 @@ class VFolder {
       recursive: recursive,
     };
     let rqst = this.client.newSignedRequest(
-      'DELETE',
-      `${this.urlPrefix}/${name}/delete_files`,
+      this.client._managerVersion >= '24.03.7' ? 'POST' : 'DELETE',
+      `${this.urlPrefix}/${name}/delete-files`,
       body,
     );
     return this.client._wrapWithPromise(rqst);
